Southern Ocean management map layers

Function for adding management layers to SOmap

SOmanagement(
  ccamlr = FALSE,
  ccamlr_labels = FALSE,
  ssru = FALSE,
  ssru_labels = FALSE,
  ssmu = FALSE,
  ssmu_labels = FALSE,
  rb = FALSE,
  rb_labels = FALSE,
  sprfmorb = FALSE,
  trim = -45,
  eez = FALSE,
  eez_labels = FALSE,
  mpa = FALSE,
  mpa_labels = FALSE,
  iwc = FALSE,
  iwc_labels = FALSE,
  domains = FALSE,
  domains_labels = FALSE,
  rb_col = "green",
  sprfmo_col = "grey50",
  ccamlr_col = "red",
  ssru_col = "grey50",
  ssmu_col = "grey70",
  eez_col = "maroon",
  mpa_col = "yellow",
  iwc_col = "blue",
  domains_col = "magenta",
  basemap
)

Arguments

ccamlr
logical: if TRUE, insert the CCAMLR area boundaries.

ccamlr_labels
logical: if TRUE, add labels for the CCAMLR areas.

ssru
logical: if TRUE, insert the CCAMLR small scale research unit boundaries.

ssru_labels
logical: if TRUE, add labels for the CCAMLR small scale research units.

ssmu
logical: if TRUE, insert the CCAMLR small scale management unit boundaries.

ssmu_labels
logical: if TRUE, add labels for the CCAMLR small scale management units.

rb
logical: if TRUE, insert the CCAMLR research block boundaries.

rb_labels
logical: if TRUE, add labels for the CCAMLR research blocks.

sprfmorb
logical: if TRUE, insert the SPRFMO toothfish research block boundaries.

trim
numeric: latitude to trim the map to. Set this to -10 for effectively no trim.

eez
logical: if TRUE, insert Exclusive Economic Zones.

eez_labels
logical: if TRUE, add labels for the Exclusive Economic Zones.

mpa
logical: if TRUE, insert CCAMLR Marine Protected Areas.

mpa_labels
logical: if TRUE, add labels for the CCAMLR Marine Protected Areas.

iwc
logical: if TRUE, insert International Whaling Commission boundaries.

iwc_labels
logical: if TRUE, add labels for the International Whaling Commission areas.

domains
logical: if TRUE, insert CCAMLR Marine Protected Areas planning domains.

domains_labels
logical: if TRUE, add labels for the CCAMLR Marine Protected Area planning domains.

rb_col
character: colour for CCAMLR research blocks.

sprfmo_col
character: colour for SPRFMO toothfish research blocks

ccamlr_col
character: colour for CCAMLR boundaries

ssru_col
character: colour for CCAMLR small scale research units.

ssmu_col
character: colour for CCAMLR small scale management units.

eez_col
character: colour for Exclusive Economic Zone boundaries.

mpa_col
character: colour for CCAMLR Marine Protected Areas.

iwc_col
character: colour for IWC boundaries.

domains_col
character: colour for the CCAMLR planning domains boundaries.

basemap
SOmap or SOmap_auto: optional map object to extract extent, projection, and other information from.

Value

An object of class "SOmap_management" containing the requested management layers. Printing or plotting this object will display those layers on the current map (note that an SOmap object needs to have been plotted first)

Examples

if (FALSE) {
  tfile <- tempfile("SOmap", fileext = ".png")
  png(tfile, width=22, height=20, units='cm', res=600)
  SOmap(trim = -45)
  SOmanagement(ccamlr = TRUE, ccamlr_labels = TRUE, trim=-45)
  dev.off()
  unlink(tfile)

  ## map with non-default latitudinal extent
  SOmap(trim = -55)
  ## either provide the same extent via 'trim'
  SOmanagement(ccamlr = TRUE, ccamlr_labels = TRUE, trim = -55)

  ## or equivalently, pass the basemap to SOmanagement
  x <- SOmap(trim = -55)
  plot(x)
  SOmanagement(ccamlr = TRUE, ccamlr_labels = TRUE, basemap = x)
}