Southern Ocean Map

Function for creating round Southern Ocean maps.

SOmap(
  bathy_legend = TRUE,
  border = TRUE,
  trim = -45,
  graticules = FALSE,
  straight = FALSE,
  land = TRUE,
  land_col = "black",
  ice = TRUE,
  ice_col = "black",
  fronts = FALSE,
  fronts_col = c("hotpink", "orchid", "plum"),
  border_col = c("white", "black"),
  border_width = 2,
  graticules_col = "grey70"
)

Arguments

bathy_legend

logical: if TRUE, insert the bathymetry legend. If bathy_legend = NULL or bathy_legend = "space", then space will be left for the legend but no legend will actually be plotted. Use this if you plan to add a legend later.

border

logical: if TRUE, insert longitude border.

trim

numeric: latitude to trim the map to. Set this to -10 for effectively no trim.

graticules

logical: if TRUE, insert graticule grid.

straight

logical: if TRUE, leave a blank space on the side for a straight legend.

land

logical: if TRUE, plot coastline.

land_col

character: colour to use for coastline.

ice

logical: if TRUE, plot ice features (ice shelves, glacier tongues, and similar).

ice_col

character: colour to use for ice features.

fronts

logical or string: if TRUE or "Orsi", plot Orsi et al., (1995) ocean fronts: Subantarctic Front, Polar Front, Southern Antarctic Circumpolar Current Front. If "Park" plot the Park & Durand (2019) fronts; Northern boundary, Subantarctic Front, Polar Front, Southern Antarctic Circumpolar Current Front and Southern Boundary.

fronts_col

character: colours for fronts.

border_col

character: colours for longitude border.

border_width

numeric: thickness (in degrees of latitude) of the border.

graticules_col

string: colour for graticule grid.

Value

An object of class "SOmap", which represents a polar-stereographic map of the southern hemisphere. Printing or plotting this object will cause it to be displayed in the current graphics device.

Examples

if (FALSE) { # \dontrun{
  tfile <- tempfile("SOmap", fileext = ".png")
  png(tfile, width = 22, height = 20, units = "cm", res = 600)
  SOmap(trim = -45, graticules = TRUE)
  dev.off()
  unlink(tfile)
  SOmap(trim = -45, graticules = TRUE)
} # }