Southern Ocean Map 2

Function for creating round Southern Ocean maps with inbuild base layers.

SOmap2(Bathleg = TRUE, land = TRUE, CCAMLR = FALSE,
  CCAMLRlab = FALSE, SSRU = FALSE, SSRUlab = FALSE, SSMU = FALSE,
  SSMUlab = FALSE, RB = FALSE, RBlab = FALSE, SPRFMORB = FALSE,
  Border = TRUE, Trim = -45, Grats = FALSE, EEZ = FALSE,
  EEZlab = FALSE, MPA = FALSE, MPAlab = FALSE, Domains = FALSE,
  Domainslab = FALSE, IWC = FALSE, IWClab = FALSE,
  straight = FALSE, fronts = FALSE, frontcols = c("hotpink",
  "orchid", "plum"), rbcol = 3, sprfmocol = "grey50", ccamlrcol = 2,
  ssrucol = "grey50", ssmucol = "grey70", eezcol = "maroon",
  mpacol = "yellow", bordercol = c("white", "black"),
  gratcol = "grey70", iwccol = "blue", domcol = "magenta")

Arguments

Bathleg
Insert the bathymetry legend.

land
Plot land boundary

CCAMLR
Insert the CCAMLR boundaries.

CCAMLRlab
Insert the CCAMLR labels.

SSRU
Insert the CCAMLR small scale research unit boundaries.

SSRUlab
Insert the CCAMLR small scale research unit labels.

SSMU
Insert the CCAMLR small scale management unit boundaries.

SSMUlab
Insert the CCAMLR small scale management unit labels.

RB
Insert the CCAMLR research block boundaries.

RBlab
Insert the CCAMLR research block labels.

SPRFMORB
Insert the SPRFMO toothfish research block boundaries.

Border
Insert longitude border.

Trim
Longitude to trim map to.

Grats
Insert graticule grid.

EEZ
Insert Exclusive Economic Zones.

EEZlab
Insert Exclusive Economic Zone labels.

MPA
Insert CCAMLE Marine Protected Areas.

MPAlab
Insert CCAMLE Marine Protected Area labels.

Domains
Insert CCAMLR Marine Protected Areas planning Domains.

Domainslab
Insert CCAMLR Marine Protected Area planning Domains labels.

IWC
Insert International Whaling Commission boundaries.

IWClab
Insert International Whaling Commission labels.

straight
Do you need a blank space on the side for a straight legend.

fronts
Plot ocean fronts: Subantarctic Front, Polar Front, Southern Antarctic Circumpolar Current Front

frontcols
colors for fronts

rbcol
Color for CCAMLR research blocks.

sprfmocol
Color for SPRFMO toothfish research blocks

ccamlrcol
Color for CCAMLR boundaries

ssrucol
Color for CCAMLR small scale research units.

ssmucol
Color for CCAMLR small scale management units.

eezcol
Color for Exclusive Economic Zone boundaries; Default is maroon.

mpacol
Color for CCAMLR Marine Protected Areas; Default is yellow.

bordercol
Colors for longitude border; Default is c("black","white").

gratcol
Color for graticule grid; Default is grey.

iwccol
Color for IWC boundaries; Default is blue.

domcol
Color for the Domain boundaries. Default is magenta.

Value

Produces at the very base a round bathymetry map of the southern hemisphere.

Examples

# NOT RUN {
SOmap2(CCAMLR = TRUE, MPA = TRUE, Trim = -45)
# }