Contextual data for Southern Ocean maps

Various spatial datasets that are commonly used on Southern Ocean maps.

data(SOmap_data)

Format

A list containing the following elements:

CCAMLR_MPA

Description: current marine protected areas

Source: CCAMLR

URL: https://data.ccamlr.org/dataset/marine-protected-areas

License: not specified

CCAMLR_research_blocks

Description: A defined spatial area in which research fishing on toothfish is conducted under a research plan agreed by the Commission

Source: CCAMLR

URL: https://data.ccamlr.org/dataset/research-blocks

License: Not specified

CCAMLR_SSMU

Description: Small-scale management units (SSMUs) are designed to be used as a basis for subdividing the precautionary catch limit for krill in Subareas 48.1, 48.2, 48.3 and 48.4, and in developing management procedures for the krill fishery that can adequately account for localised effects on krill predators (SC-CAMLR-XXI, paragraphs 3.16 to 3.18). The boundaries of the SSMUs are based on predator foraging ranges (refer SC-CAMLR-XXI, Annex 4 and Trathan et al, 2008).

Source: CCAMLR

URL: https://data.ccamlr.org/dataset/small-scale-management-units

License: Not specified

CCAMLR_SSRU

Description: Small-scale research units (SSRUs) are designed to be used as a basis for subdividing the precautionary catch limit for toothfish in exploratory fisheries, and in condcuting research fishing and developing stock assessments. The boundaries of the SSRUs are defined in Conservation Measure 41-01 (2013). CCAMLR Secretariat (2013).

Source: CCAMLR

URL: https://data.ccamlr.org/dataset/small-scale-research-units

License: Not specified

CCAMLR_statistical_areas

Description: Statistical areas, subareas and divisions are used globally for the purpose of reporting fishery statistics. CCAMLR's Convention Area in the Southern Ocean is divided, for statistical purposes, into Area 48 (Atlantic Antarctic) between 70W and 30E, Area 58 (Indian Ocean Antarctic) between 30 and 150E, and Area 88 (Pacific Antarctic) between 150E and 70W. These areas, which are further subdivided into subareas and divisions, are managed by CCAMLR.

Source: CCAMLR

URL: https://data.ccamlr.org/dataset/statistical-areas-subareas-and-divisions

License: Public domain

continent

Description: Coastline, details TBA

Source: TBA

URL: TBA

License: TBA

EEZ

Description: An exclusive economic zone (EEZ) is a sea zone prescribed by the United Nations Convention on the Law of the Sea over which a state has special rights regarding the exploration and use of marine resources. Included are the EEZs that intersect the CCAMLR convention area. Some of these are currently in dispute.

Source: CCAMLR

URL: https://data.ccamlr.org/dataset/exclusive-economic-zones

License: Not specified

fronts_orsi

Description: Southern Ocean fronts as defined by Orsi et al. 1995

Source: orsifronts

URL: https://github.com/AustralianAntarcticDivision/orsifronts

License: see orsifronts

seaice_feb and seaice_oct

Description: median October and February sea ice extent

Source: Fetterer, F., K. Knowles, W. Meier, M. Savoie, and A. K. Windnagel. 2017, updated daily. Sea Ice Index, Version 3. Boulder, Colorado USA. NSIDC: National Snow and Ice Data Center

URL: https://doi.org/10.7265/N5K072F8

License: Please cite

mirounga_leonina

Description: Example elephant seal Argos tracking data

Source: Data were sourced from the Integrated Marine Observing System (IMOS) - IMOS is supported by the Australian Government through the National Collaborative Research Infrastructure Strategy and the Super Science Initiative.

URL: https://github.com/ianjonsen/bsam

License: Please cite