Southern Ocean Map 2

Function for creating round Southern Ocean maps with inbuild base layers.

SOmap2(bathy_legend = TRUE, land = TRUE, ice = TRUE,
  ccamlr = FALSE, ccamlr_labels = FALSE, ssru = FALSE,
  ssru_labels = FALSE, ssmu = FALSE, ssmu_labels = FALSE,
  rb = FALSE, rb_labels = FALSE, sprfmorb = FALSE, border = TRUE,
  trim = -45, graticules = FALSE, eez = FALSE, eez_labels = FALSE,
  mpa = FALSE, mpa_labels = FALSE, domains = FALSE,
  domains_labels = FALSE, iwc = FALSE, iwc_labels = FALSE,
  straight = FALSE, fronts = FALSE, fronts_col = c("hotpink",
  "orchid", "plum"), land_col = "black", ice_col = "black",
  rb_col = 3, sprfmo_col = "grey50", ccamlr_col = 2,
  ssru_col = "grey50", ssmu_col = "grey70", eez_col = "maroon",
  mpa_col = "yellow", border_col = c("white", "black"),
  graticules_col = "grey70", iwc_col = "blue",
  domains_col = "magenta")

Arguments

bathy_legend
Insert the bathymetry legend.

land
Plot land boundary

ice
logical: if TRUE, plot ice features (ice shelves, glacier tongues, and similar)

ccamlr
Insert the CCAMLR boundaries.

ccamlr_labels
Insert the CCAMLR labels.

ssru
Insert the CCAMLR small scale research unit boundaries.

ssru_labels
Insert the CCAMLR small scale research unit labels.

ssmu
Insert the CCAMLR small scale management unit boundaries.

ssmu_labels
Insert the CCAMLR small scale management unit labels.

rb
Insert the CCAMLR research block boundaries.

rb_labels
Insert the CCAMLR research block labels.

sprfmorb
Insert the SPRFMO toothfish research block boundaries.

border
Insert longitude border.

trim
Longitude to trim map to.

graticules
Insert graticule grid.

eez
Insert Exclusive Economic Zones.

eez_labels
Insert Exclusive Economic Zone labels.

mpa
Insert CCAMLR Marine Protected Areas.

mpa_labels
Insert CCAMLR Marine Protected Area labels.

domains
Insert CCAMLR Marine Protected Areas planning domains.

domains_labels
Insert CCAMLR Marine Protected Area planning domains labels.

iwc
Insert International Whaling Commission boundaries.

iwc_labels
Insert International Whaling Commission labels.

straight
Do you need a blank space on the side for a straight legend.

fronts
Plot ocean fronts: Subantarctic Front, Polar Front, Southern Antarctic Circumpolar Current Front

fronts_col
colors for fronts

land_col
character: colour to use for coastline

ice_col
character: colour to use for ice features

rb_col
Color for CCAMLR research blocks.

sprfmo_col
Color for SPRFMO toothfish research blocks

ccamlr_col
Color for CCAMLR boundaries

ssru_col
Color for CCAMLR small scale research units.

ssmu_col
Color for CCAMLR small scale management units.

eez_col
Color for Exclusive Economic Zone boundaries; Default is maroon.

mpa_col
Color for CCAMLR Marine Protected Areas; Default is yellow.

border_col
Colors for longitude border; Default is c("black","white").

graticules_col
Color for graticule grid; Default is grey.

iwc_col
Color for IWC boundaries; Default is blue.

domains_col
Color for the CCAMLR planning domains boundaries. Default is magenta.

Value

An object of class "SOmap", which represents a polar-stereographic map of the southern hemisphere, with the chosen management layers added. Printing or plotting this object will cause it to be displayed in the current graphics device.

Examples

# NOT RUN {
  SOmap2(ccamlr = TRUE, mpa = TRUE, trim = -45)
# }