Southern Ocean Map 2

Function for creating round Southern Ocean maps with inbuild base layers.

SOmap2(bathy_legend = TRUE, land = TRUE, ice = TRUE,
  ccamlr = FALSE, ccamlr_labels = FALSE, ssru = FALSE,
  ssru_labels = FALSE, ssmu = FALSE, ssmu_labels = FALSE,
  rb = FALSE, rb_labels = FALSE, sprfmorb = FALSE, border = TRUE,
  trim = -45, graticules = FALSE, eez = FALSE, eez_labels = FALSE,
  mpa = FALSE, mpa_labels = FALSE, domains = FALSE,
  domains_labels = FALSE, iwc = FALSE, iwc_labels = FALSE,
  straight = FALSE, fronts = FALSE, fronts_col = c("hotpink",
  "orchid", "plum"), land_col = "black", ice_col = "black",
  rb_col = 3, sprfmo_col = "grey50", ccamlr_col = 2,
  ssru_col = "grey50", ssmu_col = "grey70", eez_col = "maroon",
  mpa_col = "yellow", border_col = c("white", "black"),
  graticules_col = "grey70", iwc_col = "blue",
  domains_col = "magenta")

Arguments

bathy_legend
logical: if TRUE, insert the bathymetry legend.

land
logical: if TRUE, plot the coastline.

ice
logical: if TRUE, plot ice features (ice shelves, glacier tongues, and similar).

ccamlr
logical: if TRUE, insert the CCAMLR area boundaries.

ccamlr_labels
logical: if TRUE, add labels for the CCAMLR areas.

ssru
logical: if TRUE, insert the CCAMLR small scale research unit boundaries.

ssru_labels
logical: if TRUE, add labels for the CCAMLR small scale research units.

ssmu
logical: if TRUE, insert the CCAMLR small scale management unit boundaries.

ssmu_labels
logical: if TRUE, add labels for the CCAMLR small scale management units.

rb
logical: if TRUE, insert the CCAMLR research block boundaries.

rb_labels
logical: if TRUE, add labels for the CCAMLR research blocks.

sprfmorb
logical: if TRUE, insert the SPRFMO toothfish research block boundaries.

border
logical: if TRUE, insert longitude border.

trim
numeric: latitude to trim the map to. Set this to -10 for effectively no trim.

graticules
logical: if TRUE, insert a graticule grid.

eez
logical: if TRUE, insert Exclusive Economic Zones.

eez_labels
logical: if TRUE, add labels for the Exclusive Economic Zones.

mpa
logical: if TRUE, insert CCAMLR Marine Protected Areas.

mpa_labels
logical: if TRUE, add labels for the CCAMLR Marine Protected Areas.

domains
logical: if TRUE, insert CCAMLR Marine Protected Areas planning domains.

domains_labels
logical: if TRUE, add labels for the CCAMLR Marine Protected Area planning domains.

iwc
logical: if TRUE, insert International Whaling Commission boundaries.

iwc_labels
logical: if TRUE, add labels for the International Whaling Commission areas.

straight
logical: if TRUE, leave a blank space on the side for a straight legend.

fronts
logical: if TRUE, plot ocean fronts: Subantarctic Front, Polar Front, Southern Antarctic Circumpolar Current Front.

fronts_col
character: colours to use for fronts.

land_col
character: colour to use for coastline.

ice_col
character: colour to use for ice features.

rb_col
character: colour for CCAMLR research blocks.

sprfmo_col
character: colour for SPRFMO toothfish research blocks

ccamlr_col
character: colour for CCAMLR boundaries

ssru_col
character: colour for CCAMLR small scale research units.

ssmu_col
character: colour for CCAMLR small scale management units.

eez_col
character: colour for Exclusive Economic Zone boundaries.

mpa_col
character: colour for CCAMLR Marine Protected Areas.

border_col
character: colours for longitude border.

graticules_col
character: colour for graticule grid.

iwc_col
character: colour for IWC boundaries.

domains_col
character: colour for the CCAMLR planning domains boundaries.

Value

An object of class "SOmap", which represents a polar-stereographic map of the southern hemisphere, with the chosen management layers added. Printing or plotting this object will cause it to be displayed in the current graphics device.

Examples

# NOT RUN {
  SOmap2(ccamlr = TRUE, mpa = TRUE, trim = -45)
# }